QUEUE: more type-checking, move public part to uv-queue.h
Make QUEUE a struct instead of an array of pointers.
This adds some minimal type-checking to QUEUE_*() macros.
Also, it makes the purpose of fields a bit clearer
(it's not that obvious that "void* active_reqs[2]" is not an arbitrary array
of void* pointers but is, in fact, a QUEUE).
Also, this aids in debugging queue-related issues a bit.

As a consequence, factor out public uv-queue.h header with QUEUE
definition from queue.h.
queue.h still contains the actual implementation.

This should not break the ABI as sizeof(QUEUE) is unchanged and
memory layout is the same too.